reorganize for readibility and clarity
