+++ js/browser.js
@@ -1,0 +1,14 @@
+function getCurrentTabUrl() {
+  return new Promise((resolve, reject) => {
+    chrome.tabs.query({ active: true, currentWindow: true }, function(tabs) {
+      if (chrome.runtime.lastError) {
+        return reject(chrome.runtime.lastError);
+      }
+      resolve(tabs[0].url);
+    });
+  });
+}
+
+export default {
+  getCurrentTabUrl,
+};
